@@ -7,7 +7,7 @@
         public string Name { get; set; }
         public string Email { get; set; }
         public string BirthDate { get; set; }
-        public string When { get; set; }
+        public string Timestamp { get; set; }
         public string Who { get; set; }
     }
 }
