@@ -1,17 +1,21 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using Equinox.Application.EventSourcedNormalizers;
 using Equinox.Application.ViewModels;
+using FluentValidation.Results;
 
 namespace Equinox.Application.Interfaces
 {
     public interface ICustomerAppService : IDisposable
     {
-        void Register(CustomerViewModel customerViewModel);
-        IEnumerable<CustomerViewModel> GetAll();
-        CustomerViewModel GetById(Guid id);
-        void Update(CustomerViewModel customerViewModel);
-        void Remove(Guid id);
-        IList<CustomerHistoryData> GetAllHistory(Guid id);
+        Task<IEnumerable<CustomerViewModel>> GetAll();
+        Task<CustomerViewModel> GetById(Guid id);
+        
+        Task<ValidationResult> Register(CustomerViewModel customerViewModel);
+        Task<ValidationResult> Update(CustomerViewModel customerViewModel);
+        Task<ValidationResult> Remove(Guid id);
+
+        Task<IList<CustomerHistoryData>> GetAllHistory(Guid id);
     }
 }
