@@ -5,7 +5,7 @@
         public override bool Equals(object obj)
         {
             var valueObject = obj as T;
-            return !ReferenceEquals(valueObject, null) && EqualsCore(valueObject);
+            return EqualsCore(valueObject);
         }
 
         protected abstract bool EqualsCore(T other);
@@ -19,10 +19,10 @@
 
         public static bool operator ==(ValueObject<T> a, ValueObject<T> b)
         {
-            if (ReferenceEquals(a, null) && ReferenceEquals(b, null))
+            if (a is null && b is null)
                 return true;
 
-            if (ReferenceEquals(a, null) || ReferenceEquals(b, null))
+            if (a is null || b is null)
                 return false;
 
             return a.Equals(b);
