@@ -1,72 +1,49 @@
 ﻿using Equinox.Application.Interfaces;
 using Equinox.Application.Services;
-using Equinox.Domain.CommandHandlers;
 using Equinox.Domain.Commands;
-using Equinox.Domain.Core.Bus;
 using Equinox.Domain.Core.Events;
-using Equinox.Domain.Core.Notifications;
-using Equinox.Domain.EventHandlers;
 using Equinox.Domain.Events;
 using Equinox.Domain.Interfaces;
 using Equinox.Infra.CrossCutting.Bus;
-using Equinox.Infra.CrossCutting.Identity.Authorization;
-using Equinox.Infra.CrossCutting.Identity.Models;
-using Equinox.Infra.CrossCutting.Identity.Services;
 using Equinox.Infra.Data.Context;
 using Equinox.Infra.Data.EventSourcing;
 using Equinox.Infra.Data.Repository;
 using Equinox.Infra.Data.Repository.EventSourcing;
-using Equinox.Infra.Data.UoW;
+using FluentValidation.Results;
 using MediatR;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
+using NetDevPack.Mediator;
 
 namespace Equinox.Infra.CrossCutting.IoC
 {
-    public class NativeInjectorBootStrapper
+    public static class NativeInjectorBootStrapper
     {
         public static void RegisterServices(IServiceCollection services)
         {
-            // ASP.NET HttpContext dependency
-            services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
-
             // Domain Bus (Mediator)
             services.AddScoped<IMediatorHandler, InMemoryBus>();
-
-            // ASP.NET Authorization Polices
-            services.AddSingleton<IAuthorizationHandler, ClaimsRequirementHandler>();
 
             // Application
             services.AddScoped<ICustomerAppService, CustomerAppService>();
 
             // Domain - Events
-            services.AddScoped<INotificationHandler<DomainNotification>, DomainNotificationHandler>();
             services.AddScoped<INotificationHandler<CustomerRegisteredEvent>, CustomerEventHandler>();
             services.AddScoped<INotificationHandler<CustomerUpdatedEvent>, CustomerEventHandler>();
             services.AddScoped<INotificationHandler<CustomerRemovedEvent>, CustomerEventHandler>();
 
             // Domain - Commands
-            services.AddScoped<IRequestHandler<RegisterNewCustomerCommand, bool>, CustomerCommandHandler>();
-            services.AddScoped<IRequestHandler<UpdateCustomerCommand, bool>, CustomerCommandHandler>();
-            services.AddScoped<IRequestHandler<RemoveCustomerCommand, bool>, CustomerCommandHandler>();
+            services.AddScoped<IRequestHandler<RegisterNewCustomerCommand, ValidationResult>, CustomerCommandHandler>();
+            services.AddScoped<IRequestHandler<UpdateCustomerCommand, ValidationResult>, CustomerCommandHandler>();
+            services.AddScoped<IRequestHandler<RemoveCustomerCommand, ValidationResult>, CustomerCommandHandler>();
 
             // Infra - Data
             services.AddScoped<ICustomerRepository, CustomerRepository>();
-            services.AddScoped<IUnitOfWork, UnitOfWork>();
             services.AddScoped<EquinoxContext>();
 
             // Infra - Data EventSourcing
-            services.AddScoped<IEventStoreRepository, EventStoreSQLRepository>();
+            services.AddScoped<IEventStoreRepository, EventStoreSqlRepository>();
             services.AddScoped<IEventStore, SqlEventStore>();
-            services.AddScoped<EventStoreSQLContext>();
-
-            // Infra - Identity Services
-            services.AddTransient<IEmailSender, AuthEmailMessageSender>();
-            services.AddTransient<ISmsSender, AuthSMSMessageSender>();
-
-            // Infra - Identity
-            services.AddScoped<IUser, AspNetUser>();
+            services.AddScoped<EventStoreSqlContext>();
         }
     }
 }
