@@ -1,39 +1,82 @@
-﻿using Equinox.Domain.Models;
+﻿using System.Linq;
+using System.Threading.Tasks;
+using Equinox.Domain.Models;
 using Equinox.Infra.Data.Mappings;
+using FluentValidation.Results;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Hosting;
+using NetDevPack.Data;
+using NetDevPack.Domain;
+using NetDevPack.Mediator;
+using NetDevPack.Messaging;
 
 namespace Equinox.Infra.Data.Context
 {
-    public class EquinoxContext : DbContext
+    public sealed class EquinoxContext : DbContext, IUnitOfWork
     {
-        private readonly IHostingEnvironment _env;
+        private readonly IMediatorHandler _mediatorHandler;
 
-        public EquinoxContext(IHostingEnvironment env)
+        public EquinoxContext(DbContextOptions<EquinoxContext> options, IMediatorHandler mediatorHandler) : base(options)
         {
-            _env = env;
+            _mediatorHandler = mediatorHandler;
+            ChangeTracker.QueryTrackingBehavior = QueryTrackingBehavior.NoTracking;
+            ChangeTracker.AutoDetectChangesEnabled = false;
         }
 
         public DbSet<Customer> Customers { get; set; }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
+            modelBuilder.Ignore<ValidationResult>();
+            modelBuilder.Ignore<Event>();
+
+            foreach (var property in modelBuilder.Model.GetEntityTypes().SelectMany(
+                e => e.GetProperties().Where(p => p.ClrType == typeof(string))))
+                property.SetColumnType("varchar(100)");
+
             modelBuilder.ApplyConfiguration(new CustomerMap());
                         
             base.OnModelCreating(modelBuilder);
         }
 
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+        public async Task<bool> Commit()
         {
-            // get the configuration from the app settings
-            var config = new ConfigurationBuilder()
-                .SetBasePath(_env.ContentRootPath)
-                .AddJsonFile("appsettings.json")
-                .Build();
-            
-            // define the database to use
-            optionsBuilder.UseSqlServer(config.GetConnectionString("DefaultConnection"));
+            // Dispatch Domain Events collection. 
+            // Choices:
+            // A) Right BEFORE committing data (EF SaveChanges) into the DB will make a single transaction including  
+            // side effects from the domain event handlers which are using the same DbContext with "InstancePerLifetimeScope" or "scoped" lifetime
+            // B) Right AFTER committing data (EF SaveChanges) into the DB will make multiple transactions. 
+            // You will need to handle eventual consistency and compensatory actions in case of failures in any of the Handlers. 
+            await _mediatorHandler.PublishDomainEvents(this).ConfigureAwait(false);
+
+            // After executing this line all the changes (from the Command Handler and Domain Event Handlers) 
+            // performed through the DbContext will be committed
+            var success = await SaveChangesAsync() > 0;
+
+            return success;
+        }
+    }
+
+    public static class MediatorExtension
+    {
+        public static async Task PublishDomainEvents<T>(this IMediatorHandler mediator, T ctx) where T : DbContext
+        {
+            var domainEntities = ctx.ChangeTracker
+                .Entries<Entity>()
+                .Where(x => x.Entity.DomainEvents != null && x.Entity.DomainEvents.Any());
+
+            var domainEvents = domainEntities
+                .SelectMany(x => x.Entity.DomainEvents)
+                .ToList();
+
+            domainEntities.ToList()
+                .ForEach(entity => entity.Entity.ClearDomainEvents());
+
+            var tasks = domainEvents
+                .Select(async (domainEvent) => {
+                    await mediator.PublishEvent(domainEvent);
+                });
+
+            await Task.WhenAll(tasks);
         }
     }
 }
