@@ -1,40 +1,21 @@
-﻿using System.IO;
-using Equinox.Domain.Core.Events;
+﻿using Equinox.Domain.Core.Events;
 using Equinox.Infra.Data.Mappings;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Hosting;
 
 
 namespace Equinox.Infra.Data.Context
 {
-    public class EventStoreSQLContext : DbContext
+    public class EventStoreSqlContext : DbContext
     {
-        public DbSet<StoredEvent> StoredEvent { get; set; }
-        private readonly IHostingEnvironment _env;
+        public EventStoreSqlContext(DbContextOptions<EventStoreSqlContext> options) : base(options) { }
 
-        public EventStoreSQLContext(IHostingEnvironment env)
-        {
-            _env = env;
-        }
+        public DbSet<StoredEvent> StoredEvent { get; set; }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
             modelBuilder.ApplyConfiguration(new StoredEventMap());
 
             base.OnModelCreating(modelBuilder);
-        }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            // get the configuration from the app settings
-            var config = new ConfigurationBuilder()
-                .SetBasePath(_env.ContentRootPath)
-                .AddJsonFile("appsettings.json")
-                .Build();
-
-            // define the database to use
-            optionsBuilder.UseSqlServer(config.GetConnectionString("DefaultConnection"));
         }
     }
 }
