@@ -1,22 +1,60 @@
-﻿using System.Linq;
+﻿using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 using Equinox.Domain.Interfaces;
 using Equinox.Domain.Models;
 using Equinox.Infra.Data.Context;
 using Microsoft.EntityFrameworkCore;
+using NetDevPack.Data;
 
 namespace Equinox.Infra.Data.Repository
 {
-    public class CustomerRepository : Repository<Customer>, ICustomerRepository
+    public class CustomerRepository : ICustomerRepository
     {
-        public CustomerRepository(EquinoxContext context)
-            : base(context)
-        {
+        protected readonly EquinoxContext Db;
+        protected readonly DbSet<Customer> DbSet;
 
+        public CustomerRepository(EquinoxContext context)
+        {
+            Db = context;
+            DbSet = Db.Set<Customer>();
         }
 
-        public Customer GetByEmail(string email)
+        public IUnitOfWork UnitOfWork => Db;
+
+        public async Task<Customer> GetById(Guid id)
         {
-            return DbSet.AsNoTracking().FirstOrDefault(c => c.Email == email);
+            return await DbSet.FindAsync(id);
+        }
+
+        public async Task<IEnumerable<Customer>> GetAll()
+        {
+            return await DbSet.ToListAsync();
+        }
+
+        public async Task<Customer> GetByEmail(string email)
+        {
+            return await DbSet.AsNoTracking().FirstOrDefaultAsync(c => c.Email == email);
+        }
+
+        public void Add(Customer customer)
+        {
+           DbSet.Add(customer);
+        }
+
+        public void Update(Customer customer)
+        {
+            DbSet.Update(customer);
+        }
+
+        public void Remove(Customer customer)
+        {
+            DbSet.Remove(customer);
+        }
+
+        public void Dispose()
+        {
+            Db.Dispose();
         }
     }
 }
