@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using Equinox.Domain.Core.Events;
 
 namespace Equinox.Infra.Data.Repository.EventSourcing
@@ -7,6 +8,6 @@ namespace Equinox.Infra.Data.Repository.EventSourcing
     public interface IEventStoreRepository : IDisposable
     {
         void Store(StoredEvent theEvent);
-        IList<StoredEvent> All(Guid aggregateId);
+        Task<IList<StoredEvent>> All(Guid aggregateId);
     }
 }
