@@ -1,11 +1,9 @@
-using System;
-
 namespace Equinox.UI.Web.Models
 {
     public class ErrorViewModel
     {
-        public string RequestId { get; set; }
-
-        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+        public int ErrorCode { get; set; }
+        public string Title { get; set; }
+        public string Message { get; set; }
     }
 }
