@@ -1,25 +1,12 @@
-﻿using System.Threading.Tasks;
-using Equinox.Domain.Core.Notifications;
-using MediatR;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 
 namespace Equinox.UI.Web.ViewComponents
 {
     public class SummaryViewComponent : ViewComponent
     {
-        private readonly DomainNotificationHandler _notifications;
-
-        public SummaryViewComponent(INotificationHandler<DomainNotification> notifications)
+        public IViewComponentResult Invoke()
         {
-            _notifications = (DomainNotificationHandler)notifications;
-        }
-
-        public async Task<IViewComponentResult> InvokeAsync()
-        {
-            var notificacoes = await Task.FromResult((_notifications.GetNotifications()));
-            notificacoes.ForEach(c => ViewData.ModelState.AddModelError(string.Empty, c.Value));
-
-            return View();
+            return View("Default");
         }
     }
 }
