@@ -2,51 +2,9 @@
     ViewData["Title"] = "Welcome";
 }
 
-<div id="myCarousel" class="carousel slide" data-ride="carousel" data-interval="6000">
-    <ol class="carousel-indicators">
-        <li data-target="#myCarousel" data-slide-to="0" class="active"></li>
-        <li data-target="#myCarousel" data-slide-to="1"></li>
-        <li data-target="#myCarousel" data-slide-to="2"></li>
-        <li data-target="#myCarousel" data-slide-to="3"></li>
-    </ol>
-    <div class="carousel-inner" role="listbox">
-        <div class="item active">
-            <img src="~/images/banner1.svg" alt="ASP.NET" class="img-responsive" />
-            <div class="carousel-caption" role="option">
-                <p>
-                </p>
-            </div>
-        </div>
-        <div class="item">
-            <img src="~/images/banner2.svg" alt="Visual Studio" class="img-responsive" />
-            <div class="carousel-caption" role="option">
-                <p>
-                </p>
-            </div>
-        </div>
-        <div class="item">
-            <img src="~/images/banner3.svg" alt="Package Management" class="img-responsive" />
-            <div class="carousel-caption" role="option">
-                <p>
-                </p>
-            </div>
-        </div>
-        <div class="item">
-            <img src="~/images/banner4.svg" alt="Microsoft Azure" class="img-responsive" />
-            <div class="carousel-caption" role="option">
-                <p>
-                </p>
-            </div>
-        </div>
-    </div>
-    <a class="left carousel-control" href="#myCarousel" role="button" data-slide="prev">
-        <span class="glyphicon glyphicon-chevron-left" aria-hidden="true"></span>
-        <span class="sr-only">Previous</span>
-    </a>
-    <a class="right carousel-control" href="#myCarousel" role="button" data-slide="next">
-        <span class="glyphicon glyphicon-chevron-right" aria-hidden="true"></span>
-        <span class="sr-only">Next</span>
-    </a>
+<div>
+    <img src="~/images/banner1.svg" alt="ASP.NET" class="img-responsive" />
+    <br /><br />
 </div>
 
 <div class="row">
@@ -62,11 +20,11 @@
     <div class="col-md-3">
         <h2>Technologies</h2>
         <ul>
-            <li>.NET Core 2.2</li>
-            <li>ASP.NET MVC Core 2.2</li>
-            <li>ASP.NET WebAPI Core 2.2</li>
-            <li>ASP.NET Identity Core 2.2</li>
-            <li>EF Core 2.2</li>
+            <li>.NET Core 3.1</li>
+            <li>ASP.NET MVC Core 3.1</li>
+            <li>ASP.NET WebAPI Core 3.1</li>
+            <li>ASP.NET Identity Core 3.1</li>
+            <li>EF Core 3.1</li>
             <li>AutoMapper</li>
             <li>FluentValidator</li>
             <li>MediatR</li>
@@ -74,7 +32,7 @@
         </ul>
     </div>
     <div class="col-md-3">
-        <h2>RoadMap for 2.0</h2>
+        <h2>RoadMap for v2</h2>
         <ul>
             <li>Booking application with Domain Model Pattern, CQRS and ES concepts</li>
             <li>ASP.NET Identity working throught WebAPI services</li>
